+++ ChangeLog.rst
@@ -3,6 +3,8 @@ HEAD
  - Authors of the mmCIF file itself (`_audit_author` category) can now be
    set by manipulating :attr:`ihm.System.authors`. (If this list is empty on
    output, the set of all citation authors is used instead, as before.)
+ - Any grants that supported the modeling can now be listed in
+   :attr:`ihm.System.grants`.
 
 0.6 - 2019-03-22
 ================